Modularize routes.ts wave 4: admin-settings + admin-logs + approvals
server/routes-modules/admin-settings.ts (97 lines)
  /api/admin/settings (load), /api/admin/settings/app +
  /api/admin/settings/app/reset, /api/admin/settings/personalization,
  /api/admin/settings/integrations, integration status probes
  (github / firewall), and managed-user CRUD.

server/routes-modules/admin-logs.ts (55 lines)
  /api/admin/logs (file logs + runtime events), /api/client-log
  (client error ingest, no auth — body bounded by Express default
  JSON limit), /api/admin/security-log.

server/routes-modules/approvals.ts (156 lines)
  /api/admin/approval-queue, /api/admin/approve/:id,
  /api/admin/reject/:id. The legacyEntryShape() and
  postApprovalConfirmationToConversation() helpers that used to be
  inline in routes.ts moved into this module.

routes.ts: 1,326 → 1,089 (-237 lines, -18%).
Combined waves 1-4: 2,101 → 1,089 (-48%).
